Child age range restriction now matches age listed in consent, fixes #981.

+++ app/models/child.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
+# Allows members to fill out surveys for by child.
 class Child < ApplicationRecord
-
   # Callbacks
   after_save :assign_child_surveys
 
@@ -11,8 +11,8 @@ class Child < ApplicationRecord
   # Scopes
 
   # Model Validation
-  validates_presence_of :first_name, :age, :user_id
-  validates_numericality_of :age, greater_than_or_equal_to: 0, less_than_or_equal_to: 8
+  validates :first_name, :age, :user_id, presence: true
+  validates :age, numericality: { greater_than_or_equal_to: 2, less_than_or_equal_to: 8 }
 
   # Model Relationships
   belongs_to :user
@@ -21,35 +21,30 @@ class Child < ApplicationRecord
   # Model Methods
 
   def sorted_answer_sessions
-    self.answer_sessions.joins(:survey).where.not(surveys: { slug: nil }).order(:locked, "surveys.name_en", :encounter)
+    answer_sessions.joins(:survey).where.not(surveys: { slug: nil }).order(:locked, 'surveys.name_en', :encounter)
   end
 
   def consented?
-    self.accepted_consent_at and self.accepted_consent_at < Time.zone.now
+    accepted_consent_at.present? && accepted_consent_at < Time.zone.now
   end
 
   private
 
-    def assign_child_surveys
-      remove_out_of_range_child_answer_sessions!
-      user_type = 'caregiver_child'
-      Survey.current.viewable.pediatric.where("surveys.child_min_age <= ? and surveys.child_max_age >= ?", self.age, self.age).joins(:survey_user_types).merge(SurveyUserType.current.where(user_type: user_type)).each do |survey|
-        unless survey.pediatric_diagnosed? && !self.diagnosed?
-          survey.encounters.where(launch_days_after_sign_up: 0).each do |encounter|
-            self.answer_sessions.where(encounter: encounter.slug, survey_id: survey.id, user_id: self.user_id).first_or_create
-          end
+  def assign_child_surveys
+    remove_out_of_range_child_answer_sessions!
+    user_type = 'caregiver_child'
+    Survey.current.viewable.pediatric.where('surveys.child_min_age <= ? and surveys.child_max_age >= ?', age, age).joins(:survey_user_types).merge(SurveyUserType.current.where(user_type: user_type)).each do |survey|
+      unless survey.pediatric_diagnosed? && !diagnosed?
+        survey.encounters.where(launch_days_after_sign_up: 0).each do |encounter|
+          answer_sessions.where(encounter: encounter.slug, survey_id: survey.id, user_id: self.user_id).first_or_create
         end
       end
     end
+  end
 
-    def remove_out_of_range_child_answer_sessions!
-      self.answer_sessions.joins(:survey).where.not("surveys.child_min_age <= ? and surveys.child_max_age >= ?", self.age, self.age).each do |answer_session|
-        if answer_session.answers.count == 0
-          answer_session.destroy
-        end
-      end
+  def remove_out_of_range_child_answer_sessions!
+    answer_sessions.joins(:survey).where.not('surveys.child_min_age <= ? and surveys.child_max_age >= ?', age, age).each do |answer_session|
+      answer_session.destroy if answer_session.answers.count == 0
     end
-
+  end
 end
-
-
